+++ src/WorkspaceManager.ts
@@ -158,6 +158,7 @@ export class WorkspaceManager implements vscode.Disposable {
       configuration.getGoogleTestGMockVerbose(),
       false,
       configuration.getTestNameLengthLimit(),
+      configuration.getStderrDecorator(),
     );
 
     this._disposables.push(
@@ -204,6 +205,9 @@ export class WorkspaceManager implements vscode.Disposable {
           }
           if (changeEvent.affects('test.testNameLengthLimit')) {
             this._shared.testNameLengthLimit = config.getTestNameLengthLimit();
+          }
+          if (changeEvent.affects('test.stderrDecorator')) {
+            this._shared.stderrDecorator = config.getStderrDecorator();
           }
           if (changeEvent.affectsAny('test.randomGeneratorSeed', 'gtest.treatGmockWarningAs', 'gtest.gmockVerbose')) {
             this._executableConfig.forEach(i => i.sendRetireAllExecutables());