+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Option to control stderr decorators `test.stderrDecorator`
+
 ## [4.16.0] - 2025-05-28
 
 ### Fixed